Extract out `TextSize` proto bits
