+++ crates/ruff_server/src/proto.rs
@@ -1,4 +1,6 @@
 mod text_edit;
 mod text_range;
+mod text_size;
 
 pub(crate) use text_range::TextRangeExt;
+pub(crate) use text_size::TextSizeExt;
